Allow configurable roll-over buffer size for large packets
Legacy DSP can produce packets up to 32kB or more. The roll-over buffer
is used when the packet is put at the end of the DMA buffer and rolls over
to the begining of buffer. To simplify the API and to go easy on the client,
OCC library copies such packet(s) into the roll-over buffer and resumes to
DMA buffer directly when well over the boundary. Keep default roll-over buffer
small for performance reasons but allow to work with DSP packets if necessary.