Make PCI-X OCC work again.
The legacy OCC firmware doesn't like it when RX is being enabled/disabled
on the fly. In needs to be enabled initially and never touched again. This
doesn't play well with late-RX-enable logic that was necessary to support
higher throughput rates - essentially allow client software to setup threads
and buffers and only enable RX when ready to process data.
The fix is to prevent tempering with RX register on PCI-X firmware.