bugfix: Prevent kernel panic when unloading driver.
In some cases, particularly with older Pre-processor2-like systems, there was
a kernel panic every time driver was unloaded. It wasn't observed in the new
DAQ systems. However the unloading routine was prematurely releasing device
handlers while the interrupt handler was still being active. There was a small
time window when the incoming interrupt would use invalid pointer. Probably
the legacy OCC fires an interrupt when being disabled which is why we see it
every time.