Fixed performance bug.
When copying from split DQ hardware queue into the unified queue, driver
would calculate the remaining length until the queue size and call that head.
When the packet does not fit into the head, remaining packet length is called
tail. tail is 0 most of the time and simple memcpy() of 0 bytes could be
optimized away by the compiler and not produce any side effects. However the
driver is using snsocb_rxcopy() function which DOES have side effects when
the copied lenght is 0 - it would always overwrite the start of the unified
queue with the packet-length-sized data from the start of the HW DQ queue.
As long as the consumer is catching up fast with the producer, this is only an
optmization issue as consumer is not reading that data anyway. And when it does,
described scenario does not take place as the tail was not 0.