Fixed data corruption on queue roll-over.
In split queues scenario (production PCI-X), the driver was wrong at
deciding when the roll-over occured and also calculating the head/tail
of the packet. During lab-testing with 24+1600 bytes packets, this would
cause 2 bad packets the roll-over happened. The same queue size is used
for hardware DQ as well as the unified buffer. So the unified buffer would
rollover couple of packets before the HW DQ would. It was a bit tricky to
spot what the problem was because of that queue misalignment.