Clear queue offset indexes during software reset.
On soft reset, TX producer and RX consumer indexes must be cleared. Not doing so
I observed silently 'stalling' incoming data. The FPGA would not report stalled
state but it would not produce any data either.
Legacy PCI-X and the new PCIe differ a bit in what needs to be done. For PCI-X
I'm just copying whatever the Windows driver is doing, even though the ODMA
thing is not completely clear what it is. Eric might be able to find out.
There are 3 other registers, one for each of IMQ, CQ, DQ that we know could be
cleared, but there's an additional ODMA clear-all register. Windows driver would
only touch the ODMA clear-all register. I'm sticking to that - it worked so far.