Enforcing 8-byte alignment into API and fixed bug on roll-over.
The API now enforces all data and packets to be 8-byte aligned. It does not
(yet) flag non-aligned packets as bad packets.
The other thing commited is the roll-over bug. When the last packet in the
buffer span over to the start of the buffer, the occlib would refuse to go
beyond the buffer boundary and always return partial packet at the end,
and keep returning partial packet.