Removed dispatcher class and make sure plugins can be instantiated from st.cmd
The dispatcher class was kinda an overhead as it would add complexity
for no good reason but to mark classes as dispatchers. Its functionality
was merged into OccPortDriver, which simplified the asyn port naming and
using from st.cmd.

The AdaraPlugin, as the first plugin available, is now instantiated from
where it should be, the st.cmd. Couple of template changes were required
to properly test PV access is still working nicely.