Introduced CmdDispatcher plugin
As part of the dispatcher cleanup, this new class appeared to distribute
only command packets. It can be used to limit number of threads needed
for each command handling plugin - those plugins need not to be non-
blocking any more as they can be connected to CmdDispatcher, which is
always non-blocking. All plugins connected to it can be run by single
thread.