Changing OCC data alignment requirement to 4bytes.
I've talked to Eric and the OCC protocol is supposed to be 8 byte aligned,
just the way Linux OCC driver was enforcing it up until now. However, the DSP
firmware violates it at least in 6.3 version with READ_CONFIG response
having 436 bytes of payload (not 8 byte aligned). We could fix the DSP firmware 6.3
but other firmwares might be affected and we might not want to fix everybody.
Interestingly enough, current OCC driver on Windows and the dcomserver were
happy so far with DSP 6.3, so neither of them is enforcing 8 byte alignment.

Since I can not find a reference document where the 8 byte alignment requirement
can be verified, I'm changing the Linux OCC driver and nED to loosen that
requirement to 4 bytes.