Adding RocPlugin
Introducing the RocPlugin with initializing mechanism only. It verifies the ROC board through its type
and version number. Reading status needs to be worked on. Configuration as well.
Plugin is based on new abstract class BaseModulePlugin which is a base for all module plugins. It adds
capability of creating outgoing packets based on connection type.