Introducing the occ_flashprog flash programmer for OCC boards.  This tool can be used to reprogram the 512Gb Micron StrataFlash device which holds the FPGA design.
Compilation is performed using the updated Makefile in the occhwtools
folder.  Usage instructions can be viewed by running ./occ_flashprog
without any parameters.

Notes:  on the first generation of OCC boards, it is necessary to
use section 2 (-s 2) to reprogram the board.  On the second generation
of OCC boards, use section 0 (-s 0).  Currently, other sections are not
used and will not program correctly.

The typical usage of this tool will be to quickly reprogram OCC boards
with firmware updates in the field, alleviating the requirement to
physically connect a laptop with a Xilinx programmer.  Typically, users
will want to use the Program option (-p), which will program and verify
the new firmware.