Added support for enabling/disabling error packets from FPGA
PCI-X and PCIe both support transforming detected link
errors into error packets. The feature can be turned on or
off dynamically, but it was not supported so far. Added driver
support, a new interface in occlib and corresponding PV in nED.

In addition, PCIe also provides 3 registers with error counters
which are by default never cleared. Software can do it
dynamically. Right now the driver is reseting the counters at
every device reset. We might change it later if found useful.