Bugfix: The split packet on buffer rollover could segfault
The mechanism to advance packets in the DMA buffer did not
obey the boundary properly. Apparently this was not the reason
to crash when using copied DMA buffer, but overflowing kernel's
DMA memory was not so graceful. Even though the copied
buffer did not complain with segfault, the data of the split
packet was garbage taken from the uninitialized memory.