Improved handling and presenting of the stall event
There was couple of things wrong with the stall handling. The
most significant was the duplicated value of RX enabled and RX
has data flags. Drivers poll() handler now uniquely flags stall
event and the library makes use of it.
In nED, PVs were added to transperently show the current state
of the stall, supporting both the OCC stall and copy thread
stall events.
BoardStatus PV was renamed to LastErr because that's really
what it is.