Reworked the simulation library to work with named pipes, now linked as shared library
The simulation library is now working with two named pipes, one for RX and one for TX
part of the communication. This gives us greater flexibility to simulate data, a smart
simulation program could play with the througput or we can just pipe file into it. It's
also possible to see outgoing data, allowing for a real DSP simulator program.

For interchangibility of two libraries, they're now linked as a shared library.