Handle odd number of bytes transfered through LVDS.
Found by Carl testing the HV communication, some of the commands
were not transmitted correctly to HV and the HV responded with "unknown command"
error. Sending odd number of bytes through LVDS link was ending the LVDS stream on
10th byte instead of 9th. The 10th character got accepted by ROC and transmitted to
HV. It wasn't until next complete command that HV looked at that overflowed
character and it decided that together with the new command did not like it.

Although existing functions were used for transmission, we didn't see the problem because
all data transmitted so far was 2 byte aligned.