Closing a file in DumpPlugin is now asynchronous
When closing a file, internal kernel buffers might need to get synchronized.
It was observed that close() could take several seconds when
closing a 5GB file which caused OCC to stall. This patch converts close()
to a non-blocking call while leaving open() and write() as blocking.

File gets open only when plugin is enabled and gets closed when plugin gets
disabled.