ROC v5.4 status response is very different from v5.2
Although the packet size is the same, fields returned in status response from ROC v5.4 are completely different
from v5.2. Unfortunately the Acquire status flag is dropped, so the software simulates that flag
when starting/stopping - but it doesn't recover if software is restarted.

Configuration fields were updated to handle baseline value parameters.