Fixed nsleep implementation for flashprog
The original nsleep implementation relied upon clock_settime, but
failed to check the return value.  This call was failing, which
caused the routine to not function as intended.  A number of
sleep functions were tested independently to determine the most
reliable way to implement a short sleep (a few uS or less).  Using
usleep(1) actually results in an average sleep of 70uS, which hurts
programming performance.  Using the TSC can give very short delays
without task switches (from 20ns on up), but can be unreliable if
the CPU speed is variable.  Using the clock_gettime while asking for
a 1ns delay gives an actual sleep of about 2uS on these systems.
For this application, that is an appropriate delay that is required
between back-to-back flash writes.  If the delay is not correct,
you will see "status reg error" messages, and the flash will not
be programmed correctly.