bugfix: Don't re-enable RX automatically after error has occurred
When the driver recognizes DMA stall or FIFO overflow condition in ISR,
it will immediately disable the RX. Data is bad or stopped anyway and
user need to recover manually. But since the register value was not
saved internally, the next time someone would try to send a packet the
RX would get re-enabled. Patch fixes that behavior by remembering
register configuration.