Potential double-unlock bugfix
During code review for SEQUOIA problem, double-unlock path was detected.
It hasn't been demonstrated so far as the kernel would assert and panic
loudly.
The condition to trigger this path would be if OCC interrupted with a
DMA stall condition but there was no data transfered to DMA memory.
Not likely but theoretically possible.