Optimize driver ISR for performance
The old ISR would first check for all error conditions which are rarely the case.
Also, it would de-assert OCC at the very end of the ISR and not leaving enough
time for the register write to get to the OCC, resulting in another ISR executing
only to find out there's nothing to serve and bailing out. This should reduce the
number of interrupts in half.

Tested at POWGEN with big one-per-pulse data packets for about a month
before commit.