Link all OCC implementations into single library. The API stays the same.
The need to use either driver or socket implementation was driving this
change. Before, each implementation was built as a standalone shared
library that had to be linked in at compile time. Now, using the same
API, the OCC library will detect which implementation is requested and
use it. It's essentially C++ class overloading but done in C with
function pointers, that get resolved when the connection is initialized.

This will allow us to not need to recompile the library every time
the socket implementation is requested. Moreover, it allows us to
have driver and socket OCC sources in nED at the same time.