Improve occ_data_wait() to not return success when no data available.
Previously the occdrv_data_wait() function could return success but
no data available in buffer. This was a potential corner case that we've
never really experienced or the application was resilient enough to
not choke on it. With some new use-cases this has become an issue.

The code has been improved to keep retrying until either some data
is available or the timeout expires. This required some careful edits
that would not impact performance too much. Its implemented as part
of endless loop now so extra care is necessary to ensure the loop
eventually exits.

Another potential bug was fixed as part of the rewrite. While we've
never seen it, code could dead-lock in case of continous interrupts
for DMA stalled event - OCC library code would keep looping until
there were such interrupts.