Added new loopback test using the occ_proxy.
The old loopback tool is obsolete as it only works with legacy packets.
Rather than try to extend it for new packets, there's now a Python
script to generate file of packets and the file is then piped through
occ_proxy. This approach allows for arbitrary packets to be generated
easily but always use the same transfer mechanism.